chore: move Regex into Support
