+++ src/Grammar/MatchPattern.php
@@ -4,7 +4,7 @@ namespace Phiki\Grammar;
 
 use Phiki\Contracts\ContainsCapturesInterface;
 use Phiki\MatchedPattern;
-use Phiki\Regex;
+use Phiki\Support\Regex;
 use Phiki\Tokenizer;
 
 class MatchPattern extends Pattern implements ContainsCapturesInterface
