+++ src/Grammar/Parser.php
@@ -23,7 +23,7 @@ use Phiki\Grammar\Injections\Scope;
 use Phiki\Grammar\Injections\Selector;
 use Phiki\Grammar\MatchPattern;
 use Phiki\Grammar\Pattern;
-use Phiki\Regex;
+use Phiki\Support\Regex;
 
 class Parser
 {
